requirements.txt
mcp[cli]
ipykernel
httpx
python-dotenv
requests